Added Ability to Link FB
I added a page to link FB. IT currently does not contain instructions on what parameters to pass explicitly, but I will updated that page in the future to make it as painless as possible. I also changed the way adding customers works (default group is now the organization's default group instead of hardcoded 0).